+++ pages/home.py
@@ -1,0 +1,8 @@
+import streamlit as st
+
+def landing_page():
+    """Simple landing page."""
+    st.title("AI Character Roleplay Engine")
+    st.write(
+        "Use the navigation menu to start chatting with your persona or edit their traits."
+    )
